requirements.txt
# Web Framework
flask>=3.0.0
gunicorn>=21.0.0
werkzeug>=3.0.0

# Machine Learning (CPU-only versions - much smaller!)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
numpy>=1.26.0
pandas>=2.1.0

# Image Processing
Pillow>=10.0.0
